chore: rename api_docs to apigen
