+++ README.md
@@ -23,7 +23,7 @@ bun run start
 This project is mostly Typescript/React/etc (as stated above). However, there is a small amount of Python code:
 - In the public/examples directory. This includes example Python code which is incorporated into the documentation, and typechecked for safety.
 - In the public/extracted-snippets directory. This includes example Python code that is automatically extracted from the mdx content, and is typechecked for safety. This should not be edited by hand.
-- In the scripts/api_docs module, which contains custom Python code for generating API docs for mirascope/mirascope and other libraries.
+- In the scripts/apigen module, which contains custom Python code for generating API docs for mirascope/mirascope and other libraries.
 
 ### Commands
 
@@ -44,11 +44,11 @@ This project is mostly Typescript/React/etc (as stated above). However, there is
 - `bun run lint:snippets` - Check if code snippets are up-to-date
 - `bun run lint:social` - Check if generated social cards (OG images) are up to date
 - `bun run generate-social` - Update generated social cards and metadata
-- `bun run lint:python` - Update and run typechecking / ruff checking on snippets and api_docs code
+- `bun run lint:python` - Update and run typechecking / ruff checking on snippets and apigen code
 - `bun run lint:format` - Check formatting with Prettier
 - `bun run fix:format` - Format all files with Prettier
 - `bun run fix:snippets` - Update extractable code snippets
-- `bun run fix:python` - Fix Python code in the scripts/api_docs module
+- `bun run fix:python` - Fix Python code in the scripts/apigen module
 
 ### Pre-commit Hooks
 
@@ -171,10 +171,22 @@ We have a Cloudflare worker that does some request pre-processing for us, e.g. t
 
 ### Python API Doc Generation
 
-Python API docs (e.g. for mirascope) are generated by the scripts/api_docs module.
+Python API docs (e.g. for mirascope) are generated by the scripts/apigen module.
 It depends on the target repo maintaining an api doc folder with markdown folders containing autodoc directives for documentation generation, and uses that folder's structure as the source of truth for the documentation structure.
 
-Api docs are autogenerated (do not modify by hand), and they can be regenerated via `bun run regenerate-api-docs`.
+Api docs are autogenerated (do not modify by hand), and they can be regenerated via `bun run apigen`. 
+
+You can regenerate specific files with:
+```bash
+# Regenerate a specific file
+bun run apigen --only llm/call.mdx
+
+# Regenerate a specific file from a specific source
+bun run apigen --only mirascope:llm/call.mdx
+
+# Regenerate a file and update metadata
+bun run apigen --only llm/call.mdx --with-meta
+```
 
 
 ## License
