+++ scripts/apigen/structure.py
